+++ R/update-app.R
@@ -23,6 +23,10 @@ updateF7App <- function(options, session = shiny::getDefaultReactiveDomain()) {
 #' \code{updateF7Entity} allows to update any Framework7 instance from the server.
 #' For each entity, the list of updatable properties may significantly vary. Please
 #' refer to the Framework7 documentation at \url{https://framework7.io/docs/}.
+#' Currently, \code{updateF7Entity} supports \link{f7Gauge},
+#' \link{f7Swiper}, \link{f7Searchbar},
+#' \link{f7PhotoBrowser}, \link{f7Popup} and
+#' \link{f7ListView}.
 #'
 #' @param id Element id.
 #' @param options Configuration list. Tightly depends on the entity.