+++ srcjs/handlers/updateEntity.js
@@ -7,7 +7,7 @@ $(function() {
     Shiny.addCustomMessageHandler('update-entity', function(message) {
     // Recover in which array is stored the given instance.
     // Uniqueness is ensured since HTML id are supposed to be unique.
-    var instanceFamily;
+    let instanceFamily;
     for (const property in app.store.state) {
       for (const e in app.store.state[property]) {
         if (e === message.id) {
@@ -15,15 +15,12 @@ $(function() {
         }
       }
     }
-
-    var oldInstance = app.store.state[instanceFamily][message.id];
-    var oldConfig = oldInstance.params;
-    var newConfig = app.utils.extend(oldConfig,  message.options);
-
+    let instance = app.store.state[instanceFamily][message.id];
+    let oldConfig = instance.params;
     // Destroy old instance
-    oldInstance.destroy();
-    // Create new config
-    var newInstance = app[instanceFamily].create(newConfig);
+    instance.destroy();
+    // Create new instance
+    let newInstance = app[instanceFamily].create(app.utils.extend(oldConfig,  message.options));
     // Update app data
     app.store.state[instanceFamily][message.id] = newInstance;
   });